add test for parseTypeBox
