+++ packages/runtime/src/components/chakra-ui/Table/Pagination.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useState } from 'react';
+import React from 'react';
 
 type TablePaginationProps = {
   pageNumber: number;
