+++ packages/runtime/src/components/chakra-ui/Table/index.ts
@@ -24,7 +24,7 @@ export default {
           ...DataPropertySchema,
         },
         {
-          name: 'marjorKey',
+          name: 'majorKey',
           ...MajorKeyPropertySchema,
         },
         {
