+++ packages/runtime/src/utils/initStateAndMethod.ts
@@ -1,50 +1,8 @@
-import {
-  ArrayKind,
-  BooleanKind,
-  IntegerKind,
-  NumberKind,
-  ObjectKind,
-  Static,
-  StringKind,
-  TSchema,
-  OptionalModifier,
-} from '@sinclair/typebox';
+import { TSchema } from '@sinclair/typebox';
 import { RuntimeApplication } from '../../../core/typings';
 import { registry } from '../registry';
 import { stateStore } from '../store';
-
-function parseTypeBox(tSchema: TSchema): Static<typeof tSchema> {
-  if (tSchema.modifier === OptionalModifier) {
-    return undefined;
-  }
-
-  switch (tSchema.kind) {
-    case StringKind:
-      return '';
-    case BooleanKind:
-      return false;
-    case ArrayKind:
-      return [];
-    case NumberKind:
-      return 0;
-    case IntegerKind:
-      return 0;
-    case ArrayKind:
-      return [];
-    case ArrayKind:
-      return [];
-
-    case ObjectKind:
-      const obj: Static<typeof tSchema> = {};
-      for (let key in tSchema.properties) {
-        obj[key] = parseTypeBox(tSchema.properties[key]);
-      }
-      return obj;
-
-    default:
-      return undefined;
-  }
-}
+import { parseTypeBox } from './parseTypeBox';
 
 export function initStateAndMethod(
   components: RuntimeApplication['spec']['components']
